+++ src/lib/sections/app/settings/settings-dialog.svelte
@@ -1,7 +1,9 @@
 <script lang="ts">
 	import { m } from '$lib/paraglide/messages';
+	import { Button } from '&/button';
 	import * as Dialog from '&/dialog';
 	import { Label } from '&/label';
+	import ImportCollectionsInput from './import-collections-input.svelte';
 	import LanguageSelect from './language-select.svelte';
 
 	interface Props {
@@ -10,6 +12,19 @@
 	}
 
 	const { isOpen, onClose }: Props = $props();
+
+	const exportSupportedExtensions: { label: string; extensionUrl: string }[] = [
+		{
+			label: 'Pinner',
+			extensionUrl:
+				'https://chromewebstore.google.com/detail/pinner-persisted-pinned-t/aabmkfbjcdbcjdjkdjjdpminiffikhpc'
+		},
+		{
+			label: 'Save Pinned Tabs',
+			extensionUrl:
+				'https://chromewebstore.google.com/detail/save-pinned-tabs/anmidgajdonkgmmilbccfefkfieajakd'
+		}
+	];
 </script>
 
 <Dialog.Root
@@ -31,6 +46,34 @@
 				{m.top_trite_shad_chop()}
 
 				<LanguageSelect />
+			</Label>
+
+			<Label class="flex w-full flex-col items-start gap-2">
+				<div class="flex flex-col gap-1">
+					{m.away_teary_capybara_surge()}
+					<span class="text-xs text-muted-foreground">
+						{m.real_zany_toucan_flow()}
+
+						{#each exportSupportedExtensions as ext, idx (ext.extensionUrl)}
+							{#if idx < exportSupportedExtensions.length - 1 && idx > 0}
+								,
+							{:else if idx === exportSupportedExtensions.length - 1}
+								&nbsp;&
+							{/if}
+
+							<Button
+								variant="link"
+								class="h-5 p-0 text-xs text-inherit"
+								href={ext.extensionUrl}
+								target="_blank"
+							>
+								{ext.label}
+							</Button>
+						{/each}
+					</span>
+				</div>
+
+				<ImportCollectionsInput />
 			</Label>
 		</div>